+++ handler/redoc/redoc_test.go
@@ -33,13 +33,7 @@ func TestGetHandler(t *testing.T) {
 
 	require.Nil(t, e)
 
-	t.Cleanup(func() {
-		ctx := s.GetContext()
-		ctx.Done()
-		s.Shutdown()
-		s.WaitAndStop()
-		webfmwk.Shutdown()
-	})
+	t.Cleanup(func() { require.Nil(t, s.ShutAndWait()) })
 
 	s.GET("/source", func(c webfmwk.Context) error {
 		return c.JSONOk(json.RawMessage(`{}`))