+++ poll_ping.go
@@ -1,25 +1,37 @@
 package webfmwk
 
 import (
+	"context"
 	"net/http"
+	"strings"
 	"time"
+
+	"github.com/burgesQ/webfmwk/v5/tls"
 )
 
-func (s *Server) getResp(uri string) (r *http.Response, e error) {
-	req, err := http.NewRequestWithContext(s.ctx, http.MethodGet, uri, http.NoBody)
+func getResp(ctx context.Context, uri string, cfg ...tls.IConfig) (*http.Response, error) {
+	if len(cfg) > 0 {
+
+		if strings.HasPrefix(uri, "http://") {
+			uri = uri[7:]
+		}
+
+		uri = "https://" + uri
+	}
+
+	req, err := http.NewRequestWithContext(ctx, http.MethodGet, uri, http.NoBody)
 	if err != nil {
-		return r, err
+		return nil, err
 	}
 
 	client := http.DefaultClient
-	r, e = client.Do(req)
 
-	return
+	return client.Do(req)
 }
 
 // pollPingEndpoint try to reach the /ping endpoint of the server
 // to then infrome that the server is up via the isReady channel
-func (s *Server) pollPingEndpoint(addr string) {
+func (s *Server) pollPingEndpoint(addr string, cfg ...tls.IConfig) {
 	var (
 		uri      = concatAddr(addr, s.meta.prefix)
 		duration = time.Millisecond * 10
@@ -30,6 +42,7 @@ func (s *Server) pollPingEndpoint(addr string) {
 	}
 
 	defer func() {
+		s.log.Infof("server is up")
 		s.isReady <- true
 	}()
 
@@ -42,17 +55,21 @@ func (s *Server) pollPingEndpoint(addr string) {
 		case <-delay.C:
 			delay.Reset(duration)
 			/* #nosec  */
-			if resp, e := s.getResp(uri); e != nil {
-				s.log.Infof("server not up (%q) ... %s", uri, e.Error())
+			resp, e := getResp(s.ctx, uri, cfg...)
+			if e != nil {
+				str := e.Error()
 
-				continue
-			} else if e = resp.Body.Close(); e != nil || resp.StatusCode != http.StatusOK {
-				s.log.Infof("unexpected status code, %s : %v", resp.StatusCode, e)
+				if strings.HasSuffix(str, "unknown authority") ||
+					strings.HasSuffix(str, "any IP SANs") {
+					return
+				}
+
+				s.log.Infof("server not up (%q): %T %v", uri, e, str)
 
 				continue
 			}
 
-			s.log.Infof("server is up")
+			resp.Body.Close()
 
 			return